+++ translation_finder/__init__.py
@@ -24,4 +24,4 @@ from .finder import Finder
 from .api import discover
 
 __all__ = ("Finder", "discover")
-__version__ = "0.2"
+__version__ = "0.3"